py-comicsviewer/requirements.txt
click
Flask
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
Pillow
PyMySQL
waitress
Werkzeug
zipp
flask_autoindex
